pyproject.toml
[project]
name = "multimodal-graphrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-text-splitters>=1.1.0",
    "langchain-google-genai",
    "langchain-experimental",
    "langchain-neo4j",
    "python-dotenv",
    "langchain-groq",
    "fastembed",
    "langchain_core",
    "hf-transfer",
    "moviepy",
    "webvtt-py",
    "tqdm",
    "qdrant-client",
]
